Add imports/exports for embedded refs as needed.
Find embedded references during the CollateImports pass.
Do this to figure out imports and exports. Then format
and store the expression for each embedded ref to pass
to the final disassembly listing pass.

This also means that we have to look up and prepare
plain label operands this way, too.

Only expression attributes and branches are left to
look up in the final disassembly listing pass.